// aiot-frontend/src/api/generated/device-service/models/device-v1-batch-upload-devices-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DeviceV1BatchUploadError } from './device-v1-batch-upload-error';

export interface DeviceV1BatchUploadDevicesResponse {
    'successCount'?: number;
    'failureCount'?: number;
    'errors'?: Array<DeviceV1BatchUploadError>;
}